Make the convenience alias test also work for system users
Previously, we ran an interactive shell that is not a login shell. In
this case, Bash sources commands from ~/.bashrc which in turn sources
commands from /etc/bashrc. The latter then sources the scripts in
/etc/profile.d. However, if the user is a system user that doesn't have
its ~/.bashrc file, then none of these is sourced.
The solution is to run an interactive login shell that always sources
/etc/profile which in turn sources the scripts in /etc/profile.d.