Change handling of docker_storage_volumegroup variable
Previously, the variable was not set by default and the role had
conditionals depending on whether it was set or not.
However, this turns out to be inconvenient for cases where this variable
is set for a group of hosts and then it becomes impossible to unset it
since Ansible has no way to unset a variable.
By setting docker_storage_volumegroup's default value to an empty string
and changing the role's conditionals to depend on whether the variable
has a non empty-string value or not, we can solve the above-mentioned
issue.