Change handling of docker_sudo_users variable
Sometimes, a system user for which one wants to enable passwordless sudo
access for running Docker doesn't exist at the time of running the role.
Therefore, running the convenience alias test is not possible.
Extend docker_sudo_users variable to a list of dictionaries where one
has to define the user's name. In addition, allow setting the 'test'
boolean which indicates whether to test the convenience alias for this
particular user or not.